Create two new modules to display project information
The modules rely on DOM manipulation to display the
project on screen.

index.js still reserved only for gathering and testing
our functions. There is, for example, an event listener
for the button, when pressed it asks for a title for
a project. The same title is also applied to our
object generated from saveToDo(), that later can recie-
ve to-do items.