Create a module responsible for landing and displaying to-do items
Now the 'infoToDo()' function also is able to gather all the
information about a task and return a paragraph listing it.

Change class name of some DOM elements. Add border to some ele-
ments. These changes improve the readability.

Give a better structure to the event listener in index.js.