Clean index.js global scope to avoid pollution
Separate module files into two folders: one for
DOM components, another for application logic.
Some components have a bit of both worlds, so their
classification was arbitrary.

Create a factory function called 'component' that
stores more general componentes of the page. The
same factory also helps to optimize the code, it
has better readability and helps us to identify
eventual problems related to specific parts.

Improve the event listener in index.js. Cut unecessary
code, optimize the rest and take advantage of the
objects generated by the factory functions. It seems
trivial, but it helps a lot when it comes to debugging.