Fix delete-create functionality bug
After delete a to-do item and try to
create a new one, a bug occurred.
Instead of appending the new item at the
end of the save, it replaced a pre-existing
to-do item.

The root of the bug was related to an index
problem. Somehow, after deleting a to-do item,
the index position of an element didn't up-
dated. We fixed that by mixing arrays and objects.