Create a module for a home button
It's basically a button that takes the
user to home page, where all the to-do
itens should be displayed.

For a while the button has no functiona-
lity attached, but we'll fix that soon.

The module is called 'button_home' and
shouldn't be confused with that old module
we deleted. If you have questions, consult
the previous commit to understand it better,