Fix a bug related to the delete button
If the user starts to delete the to-do items
from the oldest to the newest, a bug happens.
The last deleted task returns and pollute the
blank slate.

The bug was related to the index of
the to-do items, now it's fixed.

Remove lines of code that helped during the
debugging process (i.e. console.log).

Remove the delete and update button from the to-do
items when they are showing on the home screen. If
users want to delete or update a task, they
must go to the project/directory screen and do
these changes there. The home screen reflects these
changes.