Improve the forms module
Break the code into small pieces, so it's
easier to identify problems and better
delimit the components.

Create a function to facilitate setting
attributes for the DOM elements.

In the 'index.js' the button to add to-do
items now opens a modal and shows the form.
After filling out the form and clickling submit,
the to-do item is generated.