The button that opens the modal has only one functionality
Before, the referred button opened the modal and
also stored the code related to the submit event.

The submit event is now related to the submit button.

Separated this way, it's easier to indentify possible
problems and understand the code.

The input elements from forms no longer have a required
attribute, it's up to the user to decide what they
want to fill in or not.